// src/app/models/istoricSportivModel.ts
export interface IstoricSportivModel {
    numeCompetitie: string;
    dataStart: Date;
    dataFinal: Date;

    numeProba: string;
    timpTotal: string;
    timpInot: string;
    timpCiclism: string;
    timpAlergare: string;
    timpTranzit1: string;
    timpTranzit2: string;

    locPesteToti: number;
    locPerGen: number;
    locPerCatetogrie: number;

    puncte: number;
}